# Inventory-Service/src/main/resources/swagger.yaml
openapi: 3.0.3
info:
  title: BookYourShow - Inventory API
  description: REST APIs for managing inventory (movies, theatres, shows, seats) in BookYourShow Application
  version: 1.0.0
  contact:
    name: bookYourShow Team
    email: bookyourshow@company.com
...
servers:
  - url: http://localhost:8080/inventory/v1
    description: dev
paths:
...
  /shows:
    get:
      tags: [Shows]
      summary: Get shows for a movie grouped by theater
      operationId: getMovieShows
      parameters:
        - in: query
          name: movieId
          required: true
          description: Unique movie ID
          schema:
            type: string
            format: uuid
          example: "00000000-0000-4000-a000-000000000001"
        - in: query
          name: cityId
          required: true
          description: City ID to find shows in
          schema:
            type: integer
          example: 1
        - in: query
          name: langId
          required: false
          description: Selected language
          schema:
            type: integer
          example: 1
        - in: query
          name: formatId
          required: false
          description: Selected format
          schema:
            type: integer
          example: 2
        - in: query
          name: date
          required: false
          description: Date for shows. Defaults to today if not provided.
          schema:
            type: string
            format: date
          example: 2026-02-20
      responses:
        '200':
          description: List of theaters with showtimes for the movie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieShowListResponse'
        '400':
          description: Invalid request
        '404':
          description: Movie not found
        '500':
          description: Internal server error
    post:
      tags: [Shows]
      summary: Create a new show and auto-generate show seats (admin / theatre owner)
      description: >
        Creates a Show for the given movie-language-format and screen.
        On success, server-side creates one ShowSeat (status=AVAILABLE) per
        Seat of the chosen screen, in a single transaction.
      operationId: createShow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShowCreateRequest'
      responses:
        '201':
          description: Show created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShowResponse'
        '400':
          description: Invalid request (e.g., startTime in the past, missing fields)
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Movie-language-format or screen not found
        '409':
          description: Screen has an overlapping show in that time range
  /shows/{showId}:
    put:
      tags: [Shows]
      summary: Update a show (admin / theatre owner)
      description: Reschedule or re-price an existing show. Blocked if any seat is BOOKED.
      operationId: updateShow
      parameters:
        - in: path
          name: showId
          required: true
          description: Unique show ID
          schema:
            type: string
            format: uuid
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShowUpdateRequest'
      responses:
        '200':
          description: Show updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShowResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Show not found
        '409':
          description: Show has booked seats or overlaps with another show
    delete:
      tags: [Shows]
      summary: Delete a show (admin / theatre owner)
      description: Deletes a show and its show-seat rows. Blocked if any seat is BOOKED.
      operationId: deleteShow
      parameters:
        - in: path
          name: showId
          required: true
          description: Unique show ID
          schema:
            type: string
            format: uuid
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
      responses:
        '204':
          description: Show deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Show not found
        '409':
          description: Show has booked seats
  /shows/{showId}/seats:
    get:
      tags: [Shows]
      summary: Get seat layout and availability for a show
      operationId: getShowSeats
      parameters:
        - in: path
          name: showId
          required: true
          description: Unique show ID
          schema:
            type: string
            format: uuid
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
      responses:
        '200':
          description: Seat layout with availability
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShowSeats'
        '404':
          description: Show not found
        '500':
          description: Internal server error
  /shows/theatre:
    get:
      tags: [Shows]
      summary: Get all shows for a theatre grouped by movie and language-format
      operationId: getTheatreShows
      parameters:
        - in: query
          name: theatreId
          required: true
          description: Unique theatre ID
          schema:
            type: string
            format: uuid
          example: "00000000-0000-4000-b000-000000000001"
        - in: query
          name: date
          required: false
          description: Date for shows. Defaults to today if not provided.
          schema:
            type: string
            format: date
          example: 2026-03-01
      responses:
        '200':
          description: Shows grouped by movie and language-format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TheatreShowsDetailResponse'
        '404':
          description: Theatre not found
        '500':
          description: Internal server error

  # ── Internal ───────────────────────────────────────────
  /internal/shows/{showId}/seats/lock:
    post:
      tags: [Internal]
      summary: Lock seats for a show and return show details
      description: >
        Atomically verifies seat availability and marks them as BLOCKED.
        Returns show snapshot data needed by the booking service.
      operationId: lockSeats
      parameters:
        - in: path
          name: showId
          required: true
          schema:
            type: string
            format: uuid
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeatLockRequest'
      responses:
        '200':
          description: Seats locked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatLockResponse'
        '404':
          description: Show or seats not found
        '409':
          description: One or more seats are not available
  /internal/shows/{showId}/seats/release:
    post:
      tags: [Internal]
      summary: Release locked seats back to available
      operationId: releaseSeats
      parameters:
        - in: path
          name: showId
          required: true
          schema:
            type: string
            format: uuid
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeatStatusUpdateRequest'
      responses:
        '200':
          description: Seats released successfully
  /internal/shows/{showId}/seats/confirm:
    post:
      tags: [Internal]
      summary: Confirm locked seats as booked
      operationId: confirmSeats
      parameters:
        - in: path
          name: showId
          required: true
          schema:
            type: string
            format: uuid
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeatStatusUpdateRequest'
      responses:
        '200':
          description: Seats confirmed as booked

  # ── Theatres ────────────────────────────────────────────
...
components:
  schemas:
...
    MovieShowListResponse:
      type: object
      properties:
        shows:
          type: array
          items:
            $ref: '#/components/schemas/TheatreShows'
...
    Movie:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "00000000-0000-4000-a000-000000000001"
        title:
          type: string
          example: Pushpa 2
        genre:
          type: string
          example: Action
        duration:
          type: integer
          example: 180
        rating:
          type: integer
          description: "Rating as a percentage (0-100)"
          minimum: 0
          maximum: 100
          example: 84
...
    TheatreShows:
      type: object
      properties:
        theatreId:
          type: string
          format: uuid
          example: "00000000-0000-4000-b000-000000000001"
        theatreName:
          type: string
          example: PVR Cinemas
        location:
          type: string
          example: Phoenix MarketCity, Velachery
        showTimes:
          type: array
          items:
            $ref: '#/components/schemas/ShowTime'
    ShowTime:
      type: object
      properties:
        showId:
          type: string
          format: uuid
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        startTime:
          type: string
          example: "10:00 AM"
        language:
          type: string
          example: Hindi
        format:
          type: string
          example: 2D
        price:
          type: number
          format: decimal
          example: 250.00
        availableSeats:
          type: integer
          example: 120
    ShowSeats:
      type: object
      properties:
        showId:
          type: string
          format: uuid
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        movieTitle:
          type: string
          example: Pushpa 2
        theatreName:
          type: string
          example: PVR Cinemas
        screen:
          type: string
          example: Screen 3
        showTime:
          type: string
          format: local-date-time
          example: "2025-02-15T10:00:00"
        price:
          type: number
          format: decimal
          example: 250.00
        totalSeats:
          type: integer
          example: 150
        availableSeats:
          type: integer
          example: 120
        seatLayout:
          type: array
          items:
            $ref: '#/components/schemas/SeatRow'
    SeatRow:
      type: object
      properties:
        rowLabel:
          type: string
          example: "A"
        seats:
          type: array
          items:
            $ref: '#/components/schemas/Seat'
    Seat:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        seatNumber:
          type: integer
          example: 1
        status:
          type: string
          enum: [AVAILABLE, BOOKED, BLOCKED]
          example: AVAILABLE


    # ── Theatre Shows Detail ─────────────────────────────────
    TheatreShowsDetailResponse:
      type: object
      properties:
        theatreId:
          type: string
          format: uuid
          example: "00000000-0000-4000-b000-000000000001"
        theatreName:
          type: string
          example: PVR Cinemas
        location:
          type: string
          example: Phoenix MarketCity, Velachery
        movies:
          type: array
          items:
            $ref: '#/components/schemas/MovieShows'
    MovieShows:
      type: object
      properties:
        movie:
          $ref: '#/components/schemas/Movie'
        languageFormats:
          type: array
          items:
            $ref: '#/components/schemas/LanguageFormatShows'
    LanguageFormatShows:
      type: object
      properties:
        language:
          type: string
          example: Hindi
        format:
          type: string
          example: 2D
        shows:
          type: array
          items:
            $ref: '#/components/schemas/TheatreShowTime'
    TheatreShowTime:
      type: object
      properties:
        showId:
          type: string
          format: uuid
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        startTime:
          type: string
          example: "10:00 AM"
        screen:
          type: string
          example: Screen 3
        price:
          type: number
          format: decimal
          example: 250.00
        availableSeats:
          type: integer
          example: 120

    # ── Language Format ─────────────────────────────────────
...
    SeatIdentifier:
      type: object
      required:
        - rowLabel
        - seatNumber
      properties:
        rowLabel:
          type: string
          example: "A"
        seatNumber:
          type: integer
          example: 5
    SeatLockRequest:
      type: object
      required:
        - seats
      properties:
        seats:
          type: array
          items:
            $ref: '#/components/schemas/SeatIdentifier'
          example:
            - rowLabel: "A"
              seatNumber: 5
            - rowLabel: "A"
              seatNumber: 6
    SeatLockResponse:
      type: object
      properties:
        showId:
          type: string
          format: uuid
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        theatreName:
          type: string
          example: PVR Cinemas
        screenName:
          type: string
          example: Screen 3
        movieTitle:
          type: string
          example: Pushpa 2
        showTime:
          type: string
          format: local-date-time
          example: "2026-02-16T10:00:00"
        price:
          type: number
          format: decimal
          example: 250.00
        lockedSeats:
          type: array
          items:
            $ref: '#/components/schemas/LockedSeat'
    LockedSeat:
      type: object
      properties:
        showSeatId:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        rowLabel:
          type: string
          example: "A"
        seatNumber:
          type: integer
          example: 5
    SeatStatusUpdateRequest:
      type: object
      required:
        - showSeatIds
      properties:
        showSeatIds:
          type: array
          items:
            type: string
            format: uuid
          example: ["a1b2c3d4-e5f6-7890-abcd-ef1234567890"]

    # ── Admin: Cities ───────────────────────────────────────
...
    ShowCreateRequest:
      type: object
      required:
        - mlfId
        - screenId
        - startTime
        - price
      properties:
        mlfId:
          type: string
          format: uuid
          description: ID of the movie-language-format combination
          example: "00000000-0000-4000-c000-000000000001"
        screenId:
          type: string
          format: uuid
          example: "00000000-0000-4000-c000-000000000001"
        startTime:
          type: string
          format: local-date-time
          description: Show start time (must be in the future).
          example: "2026-02-20T10:00:00"
        endTime:
          type: string
          format: local-date-time
          description: Optional. If absent, derived from movie.durationMin.
          example: "2026-02-20T13:00:00"
        price:
          type: number
          format: decimal
          minimum: 0
          example: 250.00
    ShowUpdateRequest:
      type: object
      description: All fields optional. Blocked if any seat is BOOKED.
      properties:
        mlfId:
          type: string
          format: uuid
          example: "00000000-0000-4000-c000-000000000001"
        screenId:
          type: string
          format: uuid
          example: "00000000-0000-4000-c000-000000000001"
        startTime:
          type: string
          format: local-date-time
          example: "2026-02-20T10:00:00"
        endTime:
          type: string
          format: local-date-time
          example: "2026-02-20T13:00:00"
        price:
          type: number
          format: decimal
          minimum: 0
          example: 300.00
    ShowResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        mlfId:
          type: string
          format: uuid
          example: "00000000-0000-4000-c000-000000000001"
        screenId:
          type: string
          format: uuid
          example: "00000000-0000-4000-c000-000000000001"
        screenName:
          type: string
          example: Screen 3
        theatreId:
          type: string
          format: uuid
          example: "00000000-0000-4000-b000-000000000001"
        theatreName:
          type: string
          example: PVR Cinemas
        movieId:
          type: string
          format: uuid
          example: "00000000-0000-4000-a000-000000000001"
        movieTitle:
          type: string
          example: Pushpa 2
        language:
          type: string
          example: Hindi
        format:
          type: string
          example: 2D
        startTime:
          type: string
          format: local-date-time
          example: "2026-02-20T10:00:00"
        endTime:
          type: string
          format: local-date-time
          example: "2026-02-20T13:00:00"
        price:
          type: number
          format: decimal
          example: 250.00
        totalSeats:
          type: integer
          description: Number of show-seats auto-generated (one per screen seat).
          example: 150